fix: modularizado y eliminado lo innecesario
