--- docs/Modulos/alertFecha.js
@@ -1,6 +1,0 @@
-//alert de fecha
-
-    // let date = new Date().toLocaleTimeString();
-    // window.addEventListener ('load', (event) => {
-    //     window.alert(date);
-    // })
