+++ docs/index.html
@@ -98,56 +98,7 @@
           <p id="temperature-location">Cordoba, Argentina</p>
         </div>
       </div> -->
-      <button class="extend-weather">Extend weather</button>
-      <div id="carouselDark" class="carousel carousel-dark slide" data-bs-ride="carousel">
-        <ol class="carousel-indicators">
-          <li class="active" data-bs-target="#carouselDark" data-bs-slide-to="0" ></li>
-          <li data-bs-target="#carouselDark" data-bs-slide-to="1"></li>
-          <li data-bs-target="#carouselDark" data-bs-slide-to="2"></li>
-          <li data-bs-target="#carouselDark" data-bs-slide-to="3"></li>
-
-        </ol>
-        <div class="carousel-inner">
-
-          <div class="card text-bg-dark carousel-item active">
-            <div class="card">
-              <h5 class="card-title">Card title</h5>
-              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional content. This content is a little bit longer.</p>
-              <p class="card-text">Last updated 3 mins ago</p>
-            </div>
-          </div>
-          <div class="card text-bg-dark carousel-item">
-            <div class="card">
-              <h5 class="card-title">Card title</h5>
-              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional content. This content is a little bit longer.</p>
-              <p class="card-text">Last updated 3 mins ago</p>
-            </div>
-          </div>
-          <div class="card text-bg-dark carousel-item">
-            <div class="card">
-              <h5 class="card-title">Card title</h5>
-              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional content. This content is a little bit longer.</p>
-              <p class="card-text">Last updated 3 mins ago</p>
-            </div>
-          </div>
-          <div class="card text-bg-dark carousel-item">
-            <div class="card">
-              <h5 class="card-title">Card title</h5>
-              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional content. This content is a little bit longer.</p>
-              <p class="card-text">Last updated 3 mins ago</p>
-            </div>
-          </div>
-         
-        </div>
-        <a class="carousel-control-prev" href="#carouselDark" role="button"  data-bs-target="#carouselDark" data-bs-slide="prev">
-          <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-          <span class="visually-hidden">Previous</span>
-        </a>
-        <a class="carousel-control-next" href="#carouselDark" role="button" data-bs-target="#carouselDark" data-bs-slide="next">
-          <span class="carousel-control-next-icon" aria-hidden="true"></span>
-          <span class="visually-hidden">Next</span>
-        </a>
-      </div>
+     
       <!-- Who we are -->
 
 
@@ -428,6 +379,7 @@
     <!--script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js" integrity="sha384-pprn3073KE6tl6bjs2QrFaJGz5/SUsLqktiwsUTF55Jfv3qYSDhgCecCxMW52nD2" crossorigin="anonymous"></script>-->
 <!-- <script src="./app.js" type="module"></script> -->
 <script src="./main.js" type="module"></script>
+
 <script src="./Modulos/app_questions_game/gameApp.js" type="module"></script>
 
   </body>
