+++ docs/main.js
@@ -1,39 +1,12 @@
 // import { menuWithKeys } from './menuWithKeys.js';
 import { Application } from 'https://unpkg.com/@hotwired/stimulus/dist/stimulus.js';
-import { KeynavController } from '../docs/keynav_controller.js';
+import { KeyNavController } from '../docs/Modulos/keynav_controller.js';
+import { getCurrentWheather } from '../docs/Modulos/weather.js';
+import { changePositions } from '../docs/Modulos/changePositions.js';
+
 const app = Application.start();
-app.register('kn', KeynavController);
-// const event = new KeynavController();
-
-
-async function showPosition(position) {
-  lat = await position.coords.latitude;
-  long = await position.coords.longitude;
-}
-
-const getCurrentWheather = () => {
-  navigator.geolocation.getCurrentPosition(async (position) => {
-    const lat = position.coords.latitude;
-    const long = position.coords.longitude;
-
-    const reponse = await fetch(
-      `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${long}&appid=b6411eaff9ca5fbd2bbbe36b0555425b&units=metric`
-    );
-    const data = reponse.json();
-    console.log(data);
-    data.then((value) => {
-      console.log(value);
-      const wheatherSpan = document.getElementById('currentWheader');
-      const parraf = document.createElement('p');
-      parraf.id = 'parraf';
-      parraf.className = 'text-of-clima';
-      parraf.innerHTML = ` The current wheather in ${value.name} is: ${value.weather[0].description}. The temperature is ${value.main.temp} C° <span style='font-size:15px;'>&#128560;</span>. Max: ${value.main.temp_max}C°  Min : ${value.main.temp_min}C° . The humidity is ${value.main.humidity} %.  `;
-      wheatherSpan.appendChild(parraf);
-    })
-    .catch(error=> showError(error))
-  });
-};
-
+app.register('kn', KeyNavController);
 getCurrentWheather();
-
-main();
+setInterval(() => {
+  changePositions();
+}, 2000);
